+++ test/Fixture/Methods/PrivateInFinalClassRule/Success/AbstractClassExtendingAbstractClassWithProtectedMethod.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Methods\PrivateInFinalClassRule\Success;
+
+abstract class AbstractClassExtendingAbstractClassWithProtectedMethod extends AbstractClassWithProtectedMethod
+{
+}