+++ test/Fixture/Methods/PrivateInFinalClassRule/Success/FinalClassWithPublicMethod.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Methods\PrivateInFinalClassRule\Success;
+
+final class FinalClassWithPublicMethod
+{
+    public function method(): void
+    {
+    }
+}